Сделал правильную обработку сессий bd

+++ backend/core/init.sql
@@ -10,6 +10,6 @@ CREATE TABLE chats (
     id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY (START WITH 10000000000),
     last_message_text TEXT DEFAULT '_Чат создан_',
     last_message_time TIMESTAMP DEFAULT now(),
-    last_message_author INT DEFAULT 0,
+    last_message_author BIGINT DEFAULT 0,
     permissions JSONB DEFAULT '{}'::JSONB
 );
